opt-in patch
