+++ src/superqt/sliders/_generic_slider.py
@@ -19,9 +19,11 @@ So that's what `_GenericSlider` is below.
 scalar (with one handle per item), and it forms the basis of
 QRangeSlider.
 """
+import os
+import platform
 from typing import Generic, TypeVar
 
-from qtpy import QtGui
+from qtpy import QT_VERSION, QtGui
 from qtpy.QtCore import QEvent, QPoint, QPointF, QRect, Qt, Signal
 from qtpy.QtWidgets import (
     QApplication,
@@ -31,7 +33,7 @@ from qtpy.QtWidgets import (
     QStylePainter,
 )
 
-from ._range_style import MONTEREY_SLIDER_STYLES_FIX, USE_MAC_SLIDER_PATCH
+from ._range_style import MONTEREY_SLIDER_STYLES_FIX
 
 _T = TypeVar("_T")
 
@@ -43,11 +45,18 @@ SC_TICKMARKS = QStyle.SubControl.SC_SliderTickmarks
 CC_SLIDER = QStyle.ComplexControl.CC_Slider
 QOVERFLOW = 2**31 - 1
 
-
-if USE_MAC_SLIDER_PATCH:
-    _MAC_SLIDER_STYLE_PATCH = MONTEREY_SLIDER_STYLES_FIX
-else:
-    _MAC_SLIDER_STYLE_PATCH = ""
+# whether to use the MONTEREY_SLIDER_STYLES_FIX QSS hack
+# for fixing sliders on macos>=12 with QT < 6
+# https://bugreports.qt.io/browse/QTBUG-98093
+# https://github.com/napari/superqt/issues/74
+USE_MAC_SLIDER_PATCH = (
+    QT_VERSION
+    and int(QT_VERSION.split(".")[0]) < 6
+    and platform.system() == "Darwin"
+    and int(platform.mac_ver()[0].split(".", maxsplit=1)[0]) >= 12
+    and os.getenv("USE_MAC_SLIDER_PATCH", "1") != "0"
+)
+print("USE_MAC_SLIDER_PATCH", USE_MAC_SLIDER_PATCH)
 
 
 class _GenericSlider(QSlider, Generic[_T]):
@@ -87,15 +96,13 @@ class _GenericSlider(QSlider, Generic[_T]):
 
         self.setAttribute(Qt.WidgetAttribute.WA_Hover)
         self.setStyleSheet("")
+        if USE_MAC_SLIDER_PATCH:
+            self.applyMacStylePatch()
+
+    def applyMacStylePatch(self) -> str:
+        self.setStyleSheet(MONTEREY_SLIDER_STYLES_FIX)
 
     # ###############  QtOverrides  #######################
-
-    def setStyleSheet(self, styleSheet: str) -> None:
-        return super().setStyleSheet(self._patch_style(styleSheet))
-
-    def _patch_style(self, styleSheet: str) -> str:
-        # only use the stylesheet patch if a stylesheet is not already set
-        return styleSheet or _MAC_SLIDER_STYLE_PATCH
 
     def value(self) -> _T:  # type: ignore
         return self._value
@@ -286,7 +293,11 @@ class _GenericSlider(QSlider, Generic[_T]):
         if opt.tickPosition != QSlider.TickPosition.NoTicks:
             opt.subControls |= SC_TICKMARKS
         painter.drawComplexControl(CC_SLIDER, opt)
-        if USE_MAC_SLIDER_PATCH and opt.tickPosition != QSlider.TickPosition.NoTicks:
+
+        if (
+            opt.tickPosition != QSlider.TickPosition.NoTicks
+            and "MONTEREY_SLIDER_STYLES_FIX" in self.styleSheet()
+        ):
             # draw tick marks manually because they are badly behaved with style sheets
             interval = opt.tickInterval or int(self._pageStep)
             _range = self._maximum - self._minimum
