+++ src/superqt/sliders/_range_style.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import os
 import platform
 import re
 from dataclasses import dataclass, replace
@@ -21,18 +20,6 @@ from qtpy.QtWidgets import QApplication, QSlider, QStyleOptionSlider
 if TYPE_CHECKING:
     from ._generic_range_slider import _GenericRangeSlider
 
-# whether to use the MONTEREY_SLIDER_STYLES_FIX QSS hack
-# for fixing sliders on macos>=12 with QT < 6
-# https://bugreports.qt.io/browse/QTBUG-98093
-# https://github.com/napari/superqt/issues/74
-USE_MAC_SLIDER_PATCH = (
-    QT_VERSION
-    and int(QT_VERSION.split(".")[0]) < 6
-    and platform.system() == "Darwin"
-    and int(platform.mac_ver()[0].split(".", maxsplit=1)[0]) >= 12
-    and not os.getenv("DISABLE_MONTEREY_SLIDER_STYLES")
-)
-
 
 @dataclass
 class RangeSliderStyle:
@@ -49,6 +36,7 @@ class RangeSliderStyle:
     v_offset: float | None = None
     h_offset: float | None = None
     has_stylesheet: bool = False
+    _macpatch: bool = False
 
     def brush(self, opt: QStyleOptionSlider) -> QBrush:
         cg = opt.palette.currentColorGroup()
@@ -106,7 +94,8 @@ class RangeSliderStyle:
         if not self.has_stylesheet:
             tp = opt.tickPosition
             if opt.orientation == Qt.Orientation.Horizontal:
-                off += self.h_offset or SYSTEM_STYLE.h_offset or 0
+                if not self._macpatch:
+                    off += self.h_offset or SYSTEM_STYLE.h_offset or 0
             else:
                 off += self.v_offset or SYSTEM_STYLE.v_offset or 0
             if tp == QSlider.TickPosition.TicksAbove:
@@ -156,18 +145,16 @@ if QT_VERSION and int(QT_VERSION.split(".")[0]) == 6:
 
 BIG_SUR_STYLE = replace(
     CATALINA_STYLE,
-    pen_active="#DFDFDF",
     brush_active="#0A81FE",
     brush_inactive="#D5D5D5",
     brush_disabled="#E6E6E6",
     tick_offset=0,
     horizontal_thickness=4,
     vertical_thickness=4,
-    h_offset=0 if USE_MAC_SLIDER_PATCH else -2,
+    h_offset=-2,
     tick_bar_alpha=0.2,
 )
 
-MONTEREY_STYLE = replace(BIG_SUR_STYLE, h_offset=0)
 if QT_VERSION and int(QT_VERSION.split(".")[0]) == 6:
     BIG_SUR_STYLE = replace(BIG_SUR_STYLE, tick_offset=-3)
 
@@ -301,6 +288,8 @@ def update_styles_from_stylesheet(obj: _GenericRangeSlider):
 # a fix for https://bugreports.qt.io/browse/QTBUG-98093
 
 MONTEREY_SLIDER_STYLES_FIX = """
+/* MONTEREY_SLIDER_STYLES_FIX */
+
 QSlider::groove {
     background: #DFDFDF;
     border: 1px solid #DBDBDB;
@@ -334,17 +323,14 @@ QSlider::handle:pressed {
     background: #F0F0F0;
 }
 
-
 QSlider::sub-page:horizontal {
     background: #0981FE;
-    border: 1px solid #097DF7;
     border-radius: 2px;
     margin: 2px;
     height: 2px;
 }
 QSlider::add-page:vertical {
     background: #0981FE;
-    border: 1px solid #097DF7;
     border-radius: 2px;
     margin: 2px 0 6px 0;
     width: 2px;
